Enum rustc_lint_defs::BuiltinLintDiag

pub enum BuiltinLintDiag {

    AbsPathWithModule(Span),
    ProcMacroDeriveResolutionFallback {
        span: Span,
        ns: Namespace,
        ident: Ident,
    },
    MacroExpandedMacroExportsAccessedByAbsolutePaths(Span),
    ElidedLifetimesInPaths(usize, Span, bool, Span),
    UnknownCrateTypes {
        span: Span,
        candidate: Option<Symbol>,
    },
    UnusedImports {
        remove_whole_use: bool,
        num_to_remove: usize,
        remove_spans: Vec<Span>,
        test_module_span: Option<Span>,
        span_snippets: Vec<String>,
    },
    RedundantImport(Vec<(Span, bool)>, Ident),
    DeprecatedMacro {
        suggestion: Option<Symbol>,
        suggestion_span: Span,
        note: Option<Symbol>,
        path: String,
        since_kind: DeprecatedSinceKind,
    },
    MissingAbi(Span, Abi),
    UnusedDocComment(Span),
    UnusedBuiltinAttribute {
        attr_name: Symbol,
        macro_name: String,
        invoc_span: Span,
    },
    PatternsInFnsWithoutBody {
        span: Span,
        ident: Ident,
        is_foreign: bool,
    },
    LegacyDeriveHelpers(Span),
    OrPatternsBackCompat(Span, String),
    ReservedPrefix(Span, String),
    TrailingMacro(bool, Ident),
    BreakWithLabelAndLoop(Span),
    UnicodeTextFlow(Span, String),
    UnexpectedCfgName((Symbol, Span), Option<(Symbol, Span)>),
    UnexpectedCfgValue((Symbol, Span), Option<(Symbol, Span)>),
    DeprecatedWhereclauseLocation(Span, Option<(Span, String)>),
    MissingUnsafeOnExtern {
        suggestion: Span,
    },
    SingleUseLifetime {
        param_span: Span,
        deletion_span: Option<Span>,
        use_span: Option<(Span, bool)>,
        ident: Ident,
    },
    NamedArgumentUsedPositionally {
        position_sp_to_replace: Option<Span>,
        position_sp_for_msg: Option<Span>,
        named_arg_sp: Span,
        named_arg_name: String,
        is_formatting_arg: bool,
    },
    ByteSliceInPackedStructWithDerive {
        ty: String,
    },
    UnusedExternCrate {
        removal_span: Span,
    },
    ExternCrateNotIdiomatic {
        vis_span: Span,
        ident_span: Span,
    },
    AmbiguousGlobImports {
        diag: AmbiguityErrorDiag,
    },
    AmbiguousGlobReexports {
        name: String,
        namespace: String,
        first_reexport_span: Span,
        duplicate_reexport_span: Span,
    },
    HiddenGlobReexports {
        name: String,
        namespace: String,
        glob_reexport_span: Span,
        private_item_span: Span,
    },
    UnusedQualifications {
        removal_span: Span,
    },
    UnsafeAttrOutsideUnsafe {
        attribute_name_span: Span,
        sugg_spans: (Span, Span),
    },
    AssociatedConstElidedLifetime {
        elided: bool,
        span: Span,
        lifetimes_in_scope: MultiSpan,
    },
    RedundantImportVisibility {
        span: Span,
        max_vis: String,
        import_vis: String,
    },
    UnknownDiagnosticAttribute {
        span: Span,
        typo_name: Option<Symbol>,
    },
    MacroUseDeprecated,
    UnusedMacroUse,
    PrivateExternCrateReexport {
        source: Ident,
        extern_crate_span: Span,
    },
    UnusedLabel,
    MacroIsPrivate(Ident),
    UnusedMacroDefinition(Symbol),
    MacroRuleNeverUsed(usize, Symbol),
    UnstableFeature(DiagMessage),
    AvoidUsingIntelSyntax,
    AvoidUsingAttSyntax,
    IncompleteInclude,
    UnnameableTestItems,
    DuplicateMacroAttribute,
    CfgAttrNoAttributes,
    CrateTypeInCfgAttr,
    CrateNameInCfgAttr,
    MissingFragmentSpecifier,
    MetaVariableStillRepeating(MacroRulesNormalizedIdent),
    MetaVariableWrongOperator,
    DuplicateMatcherBinding,
    UnknownMacroVariable(MacroRulesNormalizedIdent),
    UnusedCrateDependency {
        extern_crate: Symbol,
        local_crate: Symbol,
    },
    WasmCAbi,
    IllFormedAttributeInput {
        suggestions: Vec<String>,
    },
    InnerAttributeUnstable {
        is_macro: bool,
    },
    OutOfScopeMacroCalls {
        path: String,
    },

}

Variants

AbsPathWithModule(Span)

ProcMacroDeriveResolutionFallback

Fields

span: Span

ns: Namespace

ident: Ident

MacroExpandedMacroExportsAccessedByAbsolutePaths(Span)

ElidedLifetimesInPaths(usize, Span, bool, Span)

UnknownCrateTypes

Fields

span: Span

candidate: Option<Symbol>

UnusedImports

Fields

remove_whole_use: bool

num_to_remove: usize

remove_spans: Vec<Span>

test_module_span: Option<Span>

span_snippets: Vec<String>

RedundantImport(Vec<(Span, bool)>, Ident)

DeprecatedMacro

Fields

suggestion: Option<Symbol>

suggestion_span: Span

note: Option<Symbol>

path: String

since_kind: DeprecatedSinceKind

MissingAbi(Span, Abi)

UnusedDocComment(Span)

UnusedBuiltinAttribute

Fields

attr_name: Symbol

macro_name: String

invoc_span: Span

PatternsInFnsWithoutBody

Fields

span: Span

ident: Ident

is_foreign: bool

LegacyDeriveHelpers(Span)

OrPatternsBackCompat(Span, String)

ReservedPrefix(Span, String)

TrailingMacro(bool, Ident)

BreakWithLabelAndLoop(Span)

UnicodeTextFlow(Span, String)

UnexpectedCfgName((Symbol, Span), Option<(Symbol, Span)>)

UnexpectedCfgValue((Symbol, Span), Option<(Symbol, Span)>)

DeprecatedWhereclauseLocation(Span, Option<(Span, String)>)

MissingUnsafeOnExtern

Fields

suggestion: Span

SingleUseLifetime

Fields

param_span: Span

Span of the parameter which declares this lifetime.

deletion_span: Option<Span>

Span of the code that should be removed when eliding this lifetime. This span should include leading or trailing comma.

use_span: Option<(Span, bool)>

Span of the single use, or None if the lifetime is never used. If true, the lifetime will be fully elided.

ident: Ident

NamedArgumentUsedPositionally

Fields

position_sp_to_replace: Option<Span>

Span where the named argument is used by position and will be replaced with the named argument name

position_sp_for_msg: Option<Span>

Span where the named argument is used by position and is used for lint messages

named_arg_sp: Span

Span where the named argument’s name is (so we know where to put the warning message)

named_arg_name: String

String containing the named arguments name

is_formatting_arg: bool

Indicates if the named argument is used as a width/precision for formatting

ByteSliceInPackedStructWithDerive

Fields

ty: String

UnusedExternCrate

Fields

removal_span: Span

ExternCrateNotIdiomatic

Fields

vis_span: Span

ident_span: Span

AmbiguousGlobImports

Fields

diag: AmbiguityErrorDiag

AmbiguousGlobReexports

Fields

name: String

The name for which collision(s) have occurred.

namespace: String

The name space for which the collision(s) occurred in.

first_reexport_span: Span

Span where the name is first re-exported.

duplicate_reexport_span: Span

Span where the same name is also re-exported.

HiddenGlobReexports

Fields

name: String

The name of the local binding which shadows the glob re-export.

namespace: String

The namespace for which the shadowing occurred in.

glob_reexport_span: Span

The glob reexport that is shadowed by the local binding.

private_item_span: Span

The local binding that shadows the glob reexport.

UnusedQualifications

Fields

removal_span: Span

The span of the unnecessarily-qualified path to remove.

UnsafeAttrOutsideUnsafe

Fields

attribute_name_span: Span

sugg_spans: (Span, Span)

AssociatedConstElidedLifetime

Fields

elided: bool

span: Span

lifetimes_in_scope: MultiSpan

RedundantImportVisibility

Fields

span: Span

max_vis: String

import_vis: String

UnknownDiagnosticAttribute

Fields

span: Span

typo_name: Option<Symbol>

MacroUseDeprecated

UnusedMacroUse

PrivateExternCrateReexport

Fields

source: Ident

extern_crate_span: Span

UnusedLabel

MacroIsPrivate(Ident)

UnusedMacroDefinition(Symbol)

MacroRuleNeverUsed(usize, Symbol)

UnstableFeature(DiagMessage)

AvoidUsingIntelSyntax

AvoidUsingAttSyntax

IncompleteInclude

UnnameableTestItems

DuplicateMacroAttribute

CfgAttrNoAttributes

CrateTypeInCfgAttr

CrateNameInCfgAttr

MissingFragmentSpecifier

MetaVariableStillRepeating(MacroRulesNormalizedIdent)

MetaVariableWrongOperator

DuplicateMatcherBinding

UnknownMacroVariable(MacroRulesNormalizedIdent)

UnusedCrateDependency

Fields

extern_crate: Symbol

local_crate: Symbol

WasmCAbi

IllFormedAttributeInput

Fields

suggestions: Vec<String>

InnerAttributeUnstable

Fields

is_macro: bool

OutOfScopeMacroCalls

Fields

path: String