Enum rustc_lint_defs::BuiltinLintDiagnostics

pub enum BuiltinLintDiagnostics {

    Normal,
    AbsPathWithModule(Span),
    ProcMacroDeriveResolutionFallback(Span),
    MacroExpandedMacroExportsAccessedByAbsolutePaths(Span),
    ElidedLifetimesInPaths(usize, Span, bool, Span),
    UnknownCrateTypes(Span, String, String),
    UnusedImports(String, Vec<(Span, String)>, Option<Span>),
    RedundantImport(Vec<(Span, bool)>, Ident),
    DeprecatedMacro(Option<Symbol>, Span),
    MissingAbi(Span, Abi),
    UnusedDocComment(Span),
    UnusedBuiltinAttribute {
        attr_name: Symbol,
        macro_name: String,
        invoc_span: Span,
    },
    PatternsInFnsWithoutBody(Span, Ident),
    LegacyDeriveHelpers(Span),
    ProcMacroBackCompat(String),
    OrPatternsBackCompat(Span, String),
    ReservedPrefix(Span),
    TrailingMacro(bool, Ident),
    BreakWithLabelAndLoop(Span),
    NamedAsmLabel(String),
    UnicodeTextFlow(Span, String),
    UnexpectedCfg((Symbol, Span), Option<(Symbol, Span)>),
    DeprecatedWhereclauseLocation(Span, String),
    SingleUseLifetime {
        param_span: Span,
        deletion_span: Span,
        use_span: Option<(Span, bool)>,
    },
    NamedArgumentUsedPositionally {
        position_sp_to_replace: Option<Span>,
        position_sp_for_msg: Option<Span>,
        named_arg_sp: Span,
        named_arg_name: String,
        is_formatting_arg: bool,
    },

}

Variants

Normal

AbsPathWithModule(Span)

ProcMacroDeriveResolutionFallback(Span)

MacroExpandedMacroExportsAccessedByAbsolutePaths(Span)

ElidedLifetimesInPaths(usize, Span, bool, Span)

UnknownCrateTypes(Span, String, String)

UnusedImports(String, Vec<(Span, String)>, Option<Span>)

RedundantImport(Vec<(Span, bool)>, Ident)

DeprecatedMacro(Option<Symbol>, Span)

MissingAbi(Span, Abi)

UnusedDocComment(Span)

UnusedBuiltinAttribute

Fields

attr_name: Symbol

macro_name: String

invoc_span: Span

PatternsInFnsWithoutBody(Span, Ident)

LegacyDeriveHelpers(Span)

ProcMacroBackCompat(String)

OrPatternsBackCompat(Span, String)

ReservedPrefix(Span)

TrailingMacro(bool, Ident)

BreakWithLabelAndLoop(Span)

NamedAsmLabel(String)

UnicodeTextFlow(Span, String)

UnexpectedCfg((Symbol, Span), Option<(Symbol, Span)>)

DeprecatedWhereclauseLocation(Span, String)

SingleUseLifetime

Fields

param_span: Span

Span of the parameter which declares this lifetime.

deletion_span: Span

Span of the code that should be removed when eliding this lifetime. This span should include leading or trailing comma.

use_span: Option<(Span, bool)>

Span of the single use, or None if the lifetime is never used. If true, the lifetime will be fully elided.

NamedArgumentUsedPositionally

Fields

position_sp_to_replace: Option<Span>

Span where the named argument is used by position and will be replaced with the named argument name

position_sp_for_msg: Option<Span>

Span where the named argument is used by position and is used for lint messages

named_arg_sp: Span

Span where the named argument’s name is (so we know where to put the warning message)

named_arg_name: String

String containing the named arguments name

is_formatting_arg: bool

Indicates if the named argument is used as a width/precision for formatting