Enum rustc_lint_defs::Level

pub enum Level {
    Allow,
    Expect(LintExpectationId),
    Warn,
    ForceWarn(Option<LintExpectationId>),
    Deny,
    Forbid,
}

Setting for how to handle a lint.

See: https://doc.rust-lang.org/rustc/lints/levels.html

Variants

Allow

The allow level will not issue any message.

Expect(LintExpectationId)

The expect level will suppress the lint message but in turn produce a message if the lint wasn’t issued in the expected scope. Expect should not be used as an initial level for a lint.

Note that this still means that the lint is enabled in this position and should be emitted, this will in turn fulfill the expectation and suppress the lint.

See RFC 2383.

The LintExpectationId is used to later link a lint emission to the actual expectation. It can be ignored in most cases.

Warn

The warn level will produce a warning if the lint was violated, however the compiler will continue with its execution.

ForceWarn(Option<LintExpectationId>)

This lint level is a special case of Warn, that can’t be overridden. This is used to ensure that a lint can’t be suppressed. This lint level can currently only be set via the console and is therefore session specific.

The LintExpectationId is intended to fulfill expectations marked via the #[expect] attribute, that will still be suppressed due to the level.

Deny

The deny level will produce an error and stop further execution after the lint pass is complete.

Forbid

Forbid is equivalent to the deny level but can’t be overwritten like the previous levels.

Implementations

impl Level

pub fn as_str(self) -> &'static str

Converts a level to a lower-case string.

pub fn from_str(x: &str) -> Option<Level>

Converts a lower-case string to a level. This will never construct the expect level as that would require a LintExpectationId

pub fn from_attr(attr: &Attribute) -> Option<Level>

Converts a symbol to a level.

pub fn to_cmd_flag(self) -> &'static str

pub fn is_error(self) -> bool

pub fn get_expectation_id(&self) -> Option<LintExpectationId>