Enum rustc_lint_defs::LintExpectationId

pub enum LintExpectationId {
    Unstable {
        attr_id: AttrId,
        lint_index: Option<u16>,
    },
    Stable {
        hir_id: HirId,
        attr_index: u16,
        lint_index: Option<u16>,
        attr_id: Option<AttrId>,
    },
}

Each lint expectation has a LintExpectationId assigned by the LintLevelsBuilder. Expected Diagnostics get the lint level Expect which stores the LintExpectationId to match it with the actual expectation later on.

The LintExpectationId has to be stable between compilations, as diagnostic instances might be loaded from cache. Lint messages can be emitted during an EarlyLintPass operating on the AST and during a LateLintPass traversing the HIR tree. The AST doesn’t have enough information to create a stable id. The LintExpectationId will instead store the AttrId defining the expectation. These LintExpectationId will be updated to use the stable [HirId] once the AST has been lowered. The transformation is done by the LintLevelsBuilder

Each lint inside the expect attribute is tracked individually, the lint_index identifies the lint inside the attribute and ensures that the IDs are unique.

The index values have a type of u16 to reduce the size of the LintExpectationId. It’s reasonable to assume that no user will define 2^16 attributes on one node or have that amount of lints listed. u16 values should therefore suffice.

Variants

Unstable

Fields

attr_id: AttrId

lint_index: Option<u16>

Used for lints emitted during the EarlyLintPass. This id is not hash stable and should not be cached.

Stable

Fields

hir_id: HirId

attr_index: u16

lint_index: Option<u16>

attr_id: Option<AttrId>

The [HirId] that the lint expectation is attached to. This id is stable and can be cached. The additional index ensures that nodes with several expectations can correctly match diagnostics to the individual expectation.

Implementations

impl LintExpectationId

pub fn is_stable(&self) -> bool

pub fn get_lint_index(&self) -> Option<u16>

pub fn set_lint_index(&mut self, new_lint_index: Option<u16>)

pub fn normalize(self) -> Self

Prepares the id for hashing. Removes references to the ast. Should only be called when the id is stable.